pyproject.toml
[tool.poetry]
name = "idpmodem"
version = "2.0.0"
description = "A library for interfacing with an Inmarsat IsatData Pro satellite IoT modem using serial AT commands."
authors = ["geoffbrucepayne <geoff.bruce-payne@inmarsat.com>"]
license = "Apache2.0"

[tool.poetry.dependencies]
python = "^3.8"
aioserial = "^1.3.0"
pyserial-asyncio = "^0.6"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
pylint = "^2.13.5"
pytest-asyncio = "^0.18.3"
pytest-mock = "^3.7.0"
asynctest = "^0.13.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
